jax cls heirarchy
